Add function support.
Adds methods to SetElement to check whether a given SetElement object is a function, is an injection, is a surjection, or is a bijection.

Adds isElementOf() method to the Element interface. Each implementation of Element can check if a set contains itself via this method `element.isElementOf(set)` as an alternative to `set.contains(element)`.